+++ src/components/MapLegend.vue
@@ -3,7 +3,20 @@
     id="map_legend_container"
     class="map-overlay"
   >
-    <p>{{ legendTitle }}</p>
+  <div id="legendHeader">
+    <div id="legendTitle">
+        <p>{{ legendTitle }}</p>
+    </div>
+    <div id="legendToggle">
+        <a id="minus" class="legendIcon">
+            <font-awesome-icon icon="minus" />
+        </a>
+        <a id="plus" class="legendIcon">
+            <font-awesome-icon icon="plus" />
+        </a>
+    </div>
+  </div>
+    <div id="legendContent"></div>
   </div>
 </template>
 
@@ -63,7 +76,7 @@
             }
 
             let legend = this.legend;
-            legend = document.getElementById('map_legend_container');
+            legend = document.getElementById('legendContent');
             // Loop through the layers to get the layer names and the colors associated with them.
             for (let index = 0; index < layers.length; index++) {
                 let layer = layers[index];
@@ -71,8 +84,9 @@
                 let item = document.createElement('div');
                 let key = document.createElement('span');
 
+                item.style.margin = "0 10px 0 0";
                 key.style.backgroundColor = color;
-                key.style.marginRight = '5px';
+                key.style.margin = '0 5px 0 0';
                 key.style.display = 'inline-block';
                 key.style.height = '10px';
                 key.style.width = '10px';
@@ -86,16 +100,20 @@
 
             // If we have monitoring locations colors and scales add them to the legend
             if (monitoringLocationLegendColors && monitoringLocationLegendScales) {
+                let dynamicText = document.createElement('p');
+                dynamicText.setAttribute('id', 'dynamicP');
+                dynamicText.appendChild(document.createTextNode('Observations at location'));
                 legend.appendChild(document.createElement('hr'));
-                legend.appendChild(document.createTextNode('observations at location'));
+                legend.appendChild(dynamicText);
                 for (let index = 0; index < monitoringLocationLegendColors.length; index++) {
                     let layer = monitoringLocationLegendScales[index];
                     let color = monitoringLocationLegendColors[index];
                     let item = document.createElement('div');
                     let key = document.createElement('span');
-
+                    
+                    item.style.margin = "0 10px 0 0";
                     key.style.backgroundColor = color;
-                    key.style.marginRight = '5px';
+                    key.style.margin = '0 5px 0 0';
                     key.style.display = 'inline-block';
                     key.style.height = '10px';
                     key.style.width = '10px';
@@ -108,24 +126,118 @@
                     legend.appendChild(item);
                 }
             }
+
+            let legendMinus = document.getElementById("minus");
+            let legendPlus = document.getElementById("plus");
+            let legendContent = document.getElementById("legendContent");
+            let legendHeader = document.getElementById("legendHeader");
+
+            let toggle = function(clicked, switchOut){
+                if(clicked.id === "minus"){
+                    legendContent.style.display = "none";
+                }else{
+                    legendContent.style.display = "block";
+                }
+                clicked.style.display = "none";
+                switchOut.style.display = "block";
+            }
+
+            legendMinus.onclick = function(){
+                toggle(legendMinus, legendPlus);
+            }
+            legendPlus.onclick = function(){
+                toggle(legendPlus, legendMinus);
+            }
         },
     }
   }
 </script>
 
 <style scoped lang="scss">
+$border: 1px solid rgb(200, 200, 200);
   .map-overlay {
     font-family: 'merriweather',serif;
-    border-style: groove;
     position: absolute;
     z-index:1000;
     bottom: 15px;
     left: 10px;
     background: rgba(255, 255, 255, 0.8);
     margin-right: 120px;
-    overflow: auto;
+    border: $border;
     border-radius: 3px;
-    padding: 5px;
   }
 
+  #legendHeader{
+      display: flex;
+      border-bottom: $border;
+  }
+
+  #legendTitle{
+      flex: 1;
+
+    p{
+        line-height: 25px;
+        padding: 0 0 0 10px;
+        margin-right: 10px;
+    }      
+  }
+
+  #legendToggle{
+      width:25px;
+      height: 25px;
+      border-left: $border;
+  }
+
+  .legendIcon{
+      display: block;
+      width: 100%;
+      height: 100%;
+      text-align: center;
+      outline: none;
+      cursor: pointer;
+        svg {
+            margin: 4px 0 0 0;
+            width: 16px;
+            height: 16px;
+        }
+  }
+
+  #minus,
+  #legendContent{
+      display: none;
+  }
+
+  #legendContent{
+      padding: 0 10px 10px 10px;
+      margin-top: 10px;
+  }
+
+  @media screen and (min-width: 600px){
+    #minus,
+    #legendContent{
+        display: block;
+    }
+    #plus{
+        display: none;
+    }
+  }
+
+</style>
+<style lang="scss">
+    .map-overlay{
+        p{
+            margin:0;
+        }
+
+        hr{
+            margin: 10px 0 10px 0;
+        }
+    }
+    #dynamicP{
+        margin-bottom: 10px;
+    }
+
+    .collapsed{
+        background: red;
+    }
 </style>